pyproject.toml
[tool.poetry]
name = "llm-backend"
version = "0.1.0"
description = ""
authors = ["Steven Mendez <rodri.stevenmendez@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
uvicorn = "0.23.2"
langserve = { extras = ["server"], version = "0.0.40" }
httpx = "0.27.0"
pydantic = "1.10.21"
tqdm = "4.66.2"
unstructured = { extras = ["all-docs"], version = "0.13.0" }
langchain-experimental = "0.0.58"
python-dotenv = "1.0.1"
langchain-openai = "0.1.6"
langchain-community = "0.0.38"
tiktoken = "0.7.0"
psycopg = { version = "3.1.19", extras = ["binary"] }
psycopg2-binary = "^2.9.9"
pgvector = "0.2.5"
sentence-transformers = "<4.0.0"
passlib = "1.7.4"
python-jose = "3.4.0"
email-validator = "2.2.0"
bcrypt = "3.2.2"

[tool.poetry.group.dev.dependencies]
langchain-cli = "0.0.18"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"